Refactor autoScheduleSection to smaller components
