+++ frontend/src/components/Pages/FrontPage/AutoScheduleSection/AutoScheduleSection.tsx
@@ -2,18 +2,16 @@ import { Button, Icon, Table, Typography } from '@equinor/eds-core-react'
 import { useLanguageContext } from 'components/Contexts/LanguageContext'
 import { useMissionDefinitionsContext } from 'components/Contexts/MissionDefinitionsContext'
 import { StyledDialog, StyledTableBody, StyledTableCell, TextAlignedButton } from 'components/Styles/StyledComponents'
-import { DaysOfWeek, parseAutoScheduledJobIds } from 'models/AutoScheduleFrequency'
-import { config } from 'config'
+import { allDays, allDaysIndexOfToday, DaysOfWeek } from 'models/AutoScheduleFrequency'
 import styled from 'styled-components'
-import { capitalizeFirstLetter, convertUTCDateToLocalDate } from 'utils/StringFormatting'
+import { capitalizeFirstLetter } from 'utils/StringFormatting'
 import { Icons } from 'utils/icons'
 import { useState } from 'react'
 import { FormCard } from 'components/Pages/MissionDefinitionPage/MissionDefinitionStyledComponents'
 import { MissionDefinitionEditDialogContent } from 'components/Pages/MissionDefinitionPage/MissionDefinitionPage'
 import { MissionDefinition } from 'models/MissionDefinition'
 import { SelectMissionsComponent } from '../MissionOverview/ScheduleMissionDialog/SelectMissionsToScheduleDialog'
-import { BackendAPICaller } from 'api/ApiCaller'
-import { Link } from 'react-router-dom'
+import { AutoScheduleMissionTableRow } from './AutoScheduleMissionTableRow'
 
 const StyledSection = styled.div`
     display: flex;
@@ -21,19 +19,9 @@ const StyledSection = styled.div`
     max-width: 960px;
     gap: 1rem;
 `
-const StyledTableRow = styled.div`
-    display: grid;
-    align-items: center;
-    gap: 1rem;
-    grid-template-columns: 100px auto 100px;
-`
 const StyledDayOverview = styled.div`
     display: grid;
     gap: 0px;
-`
-const StyledMissionButton = styled.div`
-    display: flex;
-    padding-bottom: 30px;
 `
 const StyledView = styled.div`
     display: flex;
@@ -42,6 +30,7 @@ const StyledView = styled.div`
 const StyledContent = styled.div`
     display: flex;
     flex-direction: column;
+    gap: 30px;
     align-items: end;
     @media (max-width: 600px) {
         align-items: start;
@@ -49,120 +38,12 @@ const StyledContent = styled.div`
     max-width: 960px;
 `
 
-const StyledNextAutoMission = styled.div`
-    margin-top: 30px;
-`
-
-const StyledTable = styled(Table)`
-    width: 960px;
-    margin-top: 10px;
-    border-top: 1px solid #dcdcdc;
-
-    @media (max-width: 960px) {
-        width: 95%;
-    }
-`
-
-const skipAutoScheduledMission = async (missionId: string, timeOfDay: string) => {
-    await BackendAPICaller.skipAutoScheduledMission(missionId, timeOfDay)
-}
-
-export const allDays = [
-    DaysOfWeek.Monday,
-    DaysOfWeek.Tuesday,
-    DaysOfWeek.Wednesday,
-    DaysOfWeek.Thursday,
-    DaysOfWeek.Friday,
-    DaysOfWeek.Saturday,
-    DaysOfWeek.Sunday,
-]
-
-const allDaysIndexOfToday = (convertUTCDateToLocalDate(new Date()).getDay() + 6) % 7
-
-const AutoScheduleMissionTableRow = ({
-    day,
-    time,
-    mission,
-}: {
-    day: DaysOfWeek
-    time: string
-    mission: MissionDefinition
-}) => {
-    const { TranslateText } = useLanguageContext()
-
-    enum MissionStatusType {
-        ScheduledJob = 'ScheduledJob',
-        SkippedJob = 'SkippedJob',
-        PastJob = 'PastJob',
-        FutureUnstartedJob = 'FutureUnstartedJob',
-    }
-
-    const currentDayOfTheWeek = allDays[allDaysIndexOfToday]
-
-    const selectMissionStatusType = () => {
-        if (day !== currentDayOfTheWeek) {
-            return MissionStatusType.FutureUnstartedJob
-        }
-        if (time < convertUTCDateToLocalDate(new Date()).toISOString().substring(11, 19)) {
-            return MissionStatusType.PastJob
-        }
-        if (
-            mission.autoScheduleFrequency &&
-            mission.autoScheduleFrequency.autoScheduledJobs &&
-            parseAutoScheduledJobIds(mission.autoScheduleFrequency.autoScheduledJobs)[time]
-        ) {
-            return MissionStatusType.ScheduledJob
-        }
-        return MissionStatusType.SkippedJob
-    }
-
-    const missionStatusType = selectMissionStatusType()
-
-    const typographyColor =
-        missionStatusType === MissionStatusType.SkippedJob || missionStatusType === MissionStatusType.PastJob
-            ? 'disabled'
-            : 'primary'
-
-    return (
-        <Table.Row key={mission.id + time}>
-            <Table.Cell>
-                <StyledTableRow>
-                    <Typography color={typographyColor}>{`${time.substring(0, 5)}`}</Typography>
-                    <Typography
-                        color={typographyColor}
-                        as={Link}
-                        to={`${config.FRONTEND_BASE_ROUTE}/mission-definition/${mission.id}`}
-                        link
-                    >
-                        {mission.name}
-                    </Typography>
-                    {(missionStatusType === MissionStatusType.ScheduledJob ||
-                        missionStatusType === MissionStatusType.SkippedJob) && (
-                        <Button
-                            style={{ maxWidth: '100px' }}
-                            variant="ghost"
-                            disabled={missionStatusType === MissionStatusType.SkippedJob}
-                            onClick={() => skipAutoScheduledMission(mission.id, time)}
-                        >
-                            {missionStatusType === MissionStatusType.ScheduledJob
-                                ? TranslateText('SkipAutoMission')
-                                : TranslateText('Skipped')}
-                        </Button>
-                    )}
-                </StyledTableRow>
-            </Table.Cell>
-        </Table.Row>
-    )
-}
-
-const AutoScheduleList = () => {
+const EditAutoSchedulingButton = () => {
     const { TranslateText } = useLanguageContext()
     const { missionDefinitions } = useMissionDefinitionsContext()
+
     const [dialogOpen, setDialogOpen] = useState<boolean>(false)
     const [selectedMissions, setSelectedMissions] = useState<MissionDefinition[]>([])
-
-    const autoScheduleMissionDefinitions = missionDefinitions.filter((m) => m.autoScheduleFrequency)
-    const allDaysSortedByToday = allDays.slice(allDaysIndexOfToday).concat(allDays.slice(0, allDaysIndexOfToday))
 
     const openDialog = () => {
         setDialogOpen(true)
@@ -171,76 +52,6 @@ const AutoScheduleList = () => {
         setDialogOpen(false)
         setSelectedMissions([])
     }
-
-    const DayOverview = () =>
-        allDaysSortedByToday.map((day, index) => {
-            const missionDefinitions = autoScheduleMissionDefinitions.filter((m) =>
-                m.autoScheduleFrequency!.daysOfWeek.includes(day)
-            )
-            const timeMissionPairs = missionDefinitions
-                .map((mission) =>
-                    mission.autoScheduleFrequency!.timesOfDayCET.map((time) => {
-                        return { time, mission }
-                    })
-                )
-                .flat()
-                .sort((a, b) => (a.time > b.time ? 1 : -1))
-
-            return (
-                <Table key={day}>
-                    <Table.Head>
-                        <Table.Row>
-                            <StyledTableCell>
-                                {capitalizeFirstLetter(TranslateText(day))}
-                                {index === 0 && ` (${TranslateText('today')})`}
-                            </StyledTableCell>
-                        </Table.Row>
-                    </Table.Head>
-                    <StyledTableBody>
-                        {timeMissionPairs.length > 0 ? (
-                            timeMissionPairs.map(({ time, mission }) => (
-                                <AutoScheduleMissionTableRow
-                                    key={time + mission.id}
-                                    day={day}
-                                    time={time}
-                                    mission={mission}
-                                />
-                            ))
-                        ) : (
-                            <Table.Row>
-                                <Table.Cell>
-                                    <Typography>{TranslateText('No missions')}</Typography>
-                                </Table.Cell>
-                            </Table.Row>
-                        )}
-                    </StyledTableBody>
-                </Table>
-            )
-        })
-
-    const EditAutoSchedulingButton = () => (
-        <TextAlignedButton onClick={openDialog}>
-            <Icon name={Icons.Add} size={24} />
-            {TranslateText('Edit auto scheduling')}
-        </TextAlignedButton>
-    )
-
-    const DisplayScheduledMissions = () => (
-        <>
-            {autoScheduleMissionDefinitions.length > 0 ? (
-                <>
-                    <Typography>
-                        {TranslateText('These missions will be automatically scheduled at the specified time')}
-                    </Typography>
-                    <StyledDayOverview>
-                        <DayOverview />
-                    </StyledDayOverview>
-                </>
-            ) : (
-                <Typography>{TranslateText('There are currently no automatically scheduled missions.')}</Typography>
-            )}
-        </>
-    )
 
     const UpdateAutoSchedulingDialogContent = () => (
         <>
@@ -283,122 +94,96 @@ const AutoScheduleList = () => {
             </StyledDialog.Actions>
         </>
     )
-
-    return (
-        <StyledView>
-            <StyledContent>
-                <StyledMissionButton>
-                    <EditAutoSchedulingButton />
-                </StyledMissionButton>
-                <StyledSection>
-                    <DisplayScheduledMissions />
-                </StyledSection>
-                <StyledDialog open={dialogOpen}>
-                    {selectedMissions.length === 1 ? (
-                        <UpdateAutoSchedulingDialogContent />
-                    ) : (
-                        <SelectMissionDialogContent />
-                    )}
-                </StyledDialog>
-            </StyledContent>
-        </StyledView>
-    )
-}
-
-const ShowLessOrMoreButton = ({
-    showMore,
-    setShowMore,
-}: {
-    showMore: boolean
-    setShowMore: (newShowMore: boolean) => void
-}) => {
-    const { TranslateText } = useLanguageContext()
-
-    if (showMore) {
-        return (
-            <Button
-                variant="ghost"
-                onClick={() => {
-                    setShowMore(false)
-                }}
-            >
-                <Icon name={Icons.UpChevron} size={16} />
-                {TranslateText('Show less')}
-            </Button>
-        )
-    }
-    return (
-        <Button
-            variant="ghost"
-            onClick={() => {
-                setShowMore(true)
-            }}
-        >
-            <Icon name={Icons.DownChevron} size={16} />
-            {TranslateText('Show more')}
-        </Button>
-    )
-}
-
-export const NextAutoScheduleMissionView = () => {
-    const { TranslateText } = useLanguageContext()
-    const { missionDefinitions } = useMissionDefinitionsContext()
-    const [showMore, setShowMore] = useState(false)
-
-    const autoScheduleMissionDefinitions = missionDefinitions.filter((m) => m.autoScheduleFrequency)
-    const currentDayOfTheWeek = allDays[allDaysIndexOfToday]
-
-    const missionDefinitionList = autoScheduleMissionDefinitions.filter((m) =>
-        m.autoScheduleFrequency!.daysOfWeek.includes(currentDayOfTheWeek)
-    )
-
-    const timeMissionPairs = missionDefinitionList
-        .filter((m) => m.autoScheduleFrequency?.autoScheduledJobs)
-        .flatMap((m) =>
-            m
-                .autoScheduleFrequency!.timesOfDayCET.filter(
-                    (time) => parseAutoScheduledJobIds(m.autoScheduleFrequency!.autoScheduledJobs!)[time]
-                )
-                .map((time) => ({ time, mission: m }))
-        )
-        .sort((a, b) => (a.time > b.time ? 1 : -1))
-
     return (
         <>
-            {timeMissionPairs.length > 0 && (
-                <StyledNextAutoMission>
-                    <Typography variant="h5">{TranslateText('Next auto scheduled mission for today')}</Typography>
-                    <StyledTable>
-                        <StyledTableBody>
-                            {!showMore ? (
-                                <AutoScheduleMissionTableRow
-                                    day={currentDayOfTheWeek}
-                                    time={timeMissionPairs[0].time}
-                                    mission={timeMissionPairs[0].mission}
-                                />
-                            ) : (
-                                <>
-                                    {timeMissionPairs.map(({ time, mission }) => (
-                                        <AutoScheduleMissionTableRow
-                                            key={time + mission.id}
-                                            day={currentDayOfTheWeek}
-                                            time={time}
-                                            mission={mission}
-                                        />
-                                    ))}
-                                </>
-                            )}
-                        </StyledTableBody>
-                    </StyledTable>
-                    {timeMissionPairs.length > 1 && (
-                        <ShowLessOrMoreButton showMore={showMore} setShowMore={setShowMore} />
-                    )}
-                </StyledNextAutoMission>
-            )}
+            <TextAlignedButton onClick={openDialog}>
+                <Icon name={Icons.Add} size={24} />
+                {TranslateText('Edit auto scheduling')}
+            </TextAlignedButton>
+            <StyledDialog open={dialogOpen}>
+                {selectedMissions.length === 1 ? <UpdateAutoSchedulingDialogContent /> : <SelectMissionDialogContent />}
+            </StyledDialog>
         </>
     )
 }
 
+const DayTable = ({ day, isToday }: { day: DaysOfWeek; isToday: boolean }) => {
+    const { TranslateText } = useLanguageContext()
+    const { missionDefinitions } = useMissionDefinitionsContext()
+
+    const autoScheduledMissionDefinitionsForDay = missionDefinitions.filter((m) =>
+        m.autoScheduleFrequency?.daysOfWeek.includes(day)
+    )
+    const timeMissionPairs = autoScheduledMissionDefinitionsForDay
+        .map((mission) =>
+            mission.autoScheduleFrequency!.timesOfDayCET.map((time) => {
+                return { time, mission }
+            })
+        )
+        .flat()
+        .sort((a, b) => (a.time > b.time ? 1 : -1))
+
+    return (
+        <Table key={day}>
+            <Table.Head>
+                <Table.Row>
+                    <StyledTableCell>
+                        {capitalizeFirstLetter(TranslateText(day))}
+                        {isToday && ` (${TranslateText('today')})`}
+                    </StyledTableCell>
+                </Table.Row>
+            </Table.Head>
+            <StyledTableBody>
+                {timeMissionPairs.length > 0 ? (
+                    timeMissionPairs.map(({ time, mission }) => (
+                        <AutoScheduleMissionTableRow key={time + mission.id} day={day} time={time} mission={mission} />
+                    ))
+                ) : (
+                    <Table.Row>
+                        <Table.Cell>
+                            <Typography>{TranslateText('No missions')}</Typography>
+                        </Table.Cell>
+                    </Table.Row>
+                )}
+            </StyledTableBody>
+        </Table>
+    )
+}
+
+const DayOverview = () => {
+    const { TranslateText } = useLanguageContext()
+    const { missionDefinitions } = useMissionDefinitionsContext()
+
+    const autoScheduleMissionDefinitions = missionDefinitions.filter((m) => m.autoScheduleFrequency)
+    const allDaysSortedByToday = allDays.slice(allDaysIndexOfToday).concat(allDays.slice(0, allDaysIndexOfToday))
+
+    return (
+        <StyledSection>
+            {autoScheduleMissionDefinitions.length > 0 ? (
+                <>
+                    <Typography>
+                        {TranslateText('These missions will be automatically scheduled at the specified time')}
+                    </Typography>
+                    <StyledDayOverview>
+                        {allDaysSortedByToday.map((day, index) => (
+                            <DayTable key={day} day={day} isToday={index === 0} />
+                        ))}
+                    </StyledDayOverview>
+                </>
+            ) : (
+                <Typography>{TranslateText('There are currently no automatically scheduled missions.')}</Typography>
+            )}
+        </StyledSection>
+    )
+}
+
 export const AutoScheduleSection = () => {
-    return AutoScheduleList()
+    return (
+        <StyledView>
+            <StyledContent>
+                <EditAutoSchedulingButton />
+                <DayOverview />
+            </StyledContent>
+        </StyledView>
+    )
 }
